Updated app.py
- Renamed global variable `values` to `login_signup_values` and local
variable `values` in `add_team` method to `team_info` for better
understanding and readability.
- After this commit, the server will not send information about previous
values of different input fields on the ideasboard page when and error
occurs after submission and the ideaboard page is re-rendered with the
error message. This has been done because, in a future commit, changes
to the ideasboard.html will be made which will make the error message to
be set by the javascript present in the ideasboard.html itself. Hence,
no re-rendering of the page will be required.
- Server now sends an error flag to the front-end along with an error
message or a redirect url when idea submission is done. This is done in
accordance to the change mentioned in the above mentioned bullet point.