Major bug fixes
ideasboard.html
- In a previous commit, with hash `2fb3795`, the server's respond to the
`submit` button of this page was changed. In this commit, the following
changes have been made to work with new response:
    - Added an `<h5>` just above the `submit` button to show the result
    of idea submission.
    - Changed some parts of the `submit` button's `EventListener` for
`submit` to receive messages regarding the submission of idea from the
back-end.

References.py
- Each user now has a `name` and a `teams` field in the databse now.
- Added a method `update_user` to update user detials in the database.
- fixed a bug in `update_team` and `update_idea` method.

app.py
- Modified `submit_idea` method to now add the `team_id` of the new
team, created as a result of the new idea submission, to every team
member's info in the database, under the `teams` field in users section.